# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sankalpa Odisha API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sub-components:
    get:
      tags: [SubComponents]
      summary: List visible sub-components
      operationId: listSubComponents
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: pageSize
          schema: { type: integer }
        - in: query
          name: assignedTo
          schema: { type: integer }
        - in: query
          name: announcementId
          schema: { type: integer }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/SubComponentListResponse" } } } }
        "401": { $ref: "#/components/responses/ErrorResponse" }
    post:
      tags: [SubComponents]
      summary: Add a sub-component to an announcement
      operationId: createSubComponent
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateSubComponentRequest" } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { type: object, properties: { id: { type: integer } }, required: [id] } } } }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
...
  /comments:
    get:
      tags: [Comments]
      summary: List comments visible to user
      operationId: listComments
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: pageSize
          schema: { type: integer }
        - in: query
          name: announcementId
          schema: { type: integer }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/CommentListResponse" } } } }
        "401": { $ref: "#/components/responses/ErrorResponse" }
  /notifications:
    get:
      tags: [Notifications]
      summary: List notifications for the current user
      operationId: listNotifications
      parameters:
        - in: query
          name: unreadOnly
          required: false
          schema: { type: boolean }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: pageSize
          schema: { type: integer }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/NotificationListResponse" } } } }
        "401": { $ref: "#/components/responses/ErrorResponse" }
...
components:
  responses:
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    SubComponentStatus:
      type: string
      enum: [Pending, "In Progress", Completed]
    NotificationType:
      type: string
      enum: [info, success, warning]
...
    SubComponent:
      type: object
      properties:
        id: { type: integer }
        announcementId: { type: integer }
        title: { type: string }
        description: { type: string }
        assignedTo: { type: integer, nullable: true }
        assignedToName: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/SubComponentStatus" }
      required: [id, announcementId, title, description, status]
    SubComponentListResponse:
      type: object
      properties:
        subComponents:
          type: array
          items: { $ref: "#/components/schemas/SubComponent" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
      required: [subComponents, total, page, pageSize]
    CreateSubComponentRequest:
      type: object
      properties:
        announcementId: { type: integer }
        title: { type: string }
        description: { type: string }
        assignedTo: { type: integer, nullable: true }
        status: { $ref: "#/components/schemas/SubComponentStatus" }
      required: [announcementId, title]
...
    AnnouncementComment:
      type: object
      properties:
        id: { type: integer }
        announcementId: { type: integer }
        userId: { type: integer }
        userName: { type: string, nullable: true }
        role: { type: string, nullable: true }
        action: { type: string }
        comment: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, announcementId, userId, action, comment, createdAt]
    CommentListResponse:
      type: object
      properties:
        comments:
          type: array
          items: { $ref: "#/components/schemas/AnnouncementComment" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
      required: [comments, total, page, pageSize]
    Notification:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        message: { type: string }
        type: { $ref: "#/components/schemas/NotificationType" }
        read: { type: boolean }
        link: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [id, userId, message, type, read, createdAt]
    NotificationListResponse:
      type: object
      properties:
        notifications:
          type: array
          items: { $ref: "#/components/schemas/Notification" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
        unreadCount: { type: integer }
      required: [notifications, total, page, pageSize, unreadCount]
...